[REFACTOR] View 코드 리팩터링
