+++ POMPOM/POMPOM/UI/ClothPickerView/ViewModel/PickerViewModel.swift
@@ -8,16 +8,14 @@
 import SwiftUI
 import SystemConfiguration
 
-class PickerViewModel: ObservableObject {
+class PickerViewModel: ClothViewModel {
     //MARK: - Propeties
     @Published var currentType: ClothCategory = .hat
     //UI 에 보여지는 컬러, 옷
     @Published var currentPresets: [String] = []
     @Published var currentItems: [String] = []
     //선택된 옷 + 컬러
-    @Published var selectedItems: [ClothCategory : Cloth] = [:]
-    
-    var networkManager: ClothesManager = ClothesManager()
+
     
     
     // 기본 컬러 프리셋
@@ -58,10 +56,6 @@ class PickerViewModel: ObservableObject {
         }
     }
 
-    //MARK: - LifeCycle
-    init() {
-        changeCategory(with: .hat)
-    }
     
     //MARK: - Methods
     func changeCategory(with category: ClothCategory) {
@@ -106,30 +100,6 @@ class PickerViewModel: ObservableObject {
         } else {
             print("DEBUG: 사용자 코드 조회 실패")
         }
-    }
-    
-    //CouplleView
-    
-    func requestClothes() async {
-        if let defaultCode: String = UserDefaults.standard.string(forKey: "code") {
-            networkManager.loadClothes(userCode: defaultCode) { clothes in
-                    self.selectedItems = clothes
-            }
-        } else {
-            print("DEBUG: 사용자 코드 조회 실패")
-        }
-    }
-    
-    func clearSelectedItem() {
-        selectedItems.removeAll()
-    }
-    
-    func fetchImageString(with category: ClothCategory) -> String {
-        if let name = selectedItems[category]?.id {
-            let imageName = "\(category)-\(name)"
-            return imageName
-        }
-        return ""
     }
 }
