--- POMPOM/POMPOM/UI/CoupleView/ViewModel/CoupleViewModel.swift
@@ -1,9 +1,0 @@
-//
-//  CoupleViewModel.swift
-//  POMPOM
-//
-//  Created by GOngTAE on 2022/06/09.
-//
-
-import Foundation
-
